requirements.txt
python-dotenv~=1.0.0
openai==1.3.8
urllib3==1.26.18
requests~=2.31.0
duckduckgo-search==3.9.11
google-search-results==2.4.2
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9